Move services into proper folder
