+++ iam-login-service/src/main/java/it/infn/mw/iam/authn/oidc/service/OIDCProviderMetadataService.java
@@ -13,13 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package it.infn.mw.iam.authn.oidc;
+package it.infn.mw.iam.authn.oidc.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
 import com.fasterxml.jackson.databind.JsonNode;
 
+import it.infn.mw.iam.authn.oidc.OIDCProviderMetadata;
 import it.infn.mw.iam.core.oauth.discovery.DefaultOidcDiscoveryService;
 
 @Service
