+++ iam-login-service/src/main/java/it/infn/mw/iam/config/CacheConfig.java
@@ -28,7 +28,7 @@ import org.springframework.data.redis.cache.RedisCacheConfiguration;
 
 import com.github.benmanes.caffeine.cache.Caffeine;
 
-import it.infn.mw.iam.authn.oidc.UserInfoFetcher;
+import it.infn.mw.iam.authn.oidc.service.UserInfoFetcher;
 import it.infn.mw.iam.core.jwk.IamJWTSigningService;
 import it.infn.mw.iam.core.oauth.discovery.DefaultOidcDiscoveryService;
 import it.infn.mw.iam.core.oauth.scope.matchers.DefaultScopeMatcherRegistry;
