+++ iam-login-service/src/main/java/it/infn/mw/iam/config/MitreServicesConfig.java
@@ -237,5 +237,4 @@ public class MitreServicesConfig {
   SelfAssertionValidator selfAssertionValidator() {
     return new SelfAssertionValidator();
   }
-
 }
