+++ iam-login-service/src/main/java/it/infn/mw/iam/rcauth/DefaultRCAuthTokenRequestor.java
@@ -39,9 +39,9 @@ import org.springframework.web.client.RestTemplate;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import it.infn.mw.iam.authn.oidc.OIDCProviderMetadata;
-import it.infn.mw.iam.authn.oidc.OIDCProviderMetadataService;
 import it.infn.mw.iam.authn.oidc.RestTemplateFactory;
 import it.infn.mw.iam.authn.oidc.model.TokenEndpointErrorResponse;
+import it.infn.mw.iam.authn.oidc.service.OIDCProviderMetadataService;
 import it.infn.mw.iam.config.IamProperties;
 import it.infn.mw.iam.rcauth.oidc.RCAuthTokenResponseVerifier;
 import it.infn.mw.iam.rcauth.util.AddContentTypeInterceptor;
